fixed [ ] bug in loop closure version

+++ InfiniTAM/ITMLib/Engines/MultiScene/ITMMapGraphManager.tpp
@@ -119,8 +119,8 @@ namespace ITMLib
 	ORUtils::SE3Pose ITMVoxelMapGraphManager<TVoxel, TIndex>::findTransformation(int fromLocalMapId, int toLocalMapId) const
 	{
 		ORUtils::SE3Pose fromLocalMapPose, toLocalMapPose;
-		if ((fromLocalMapId >= 0) || (fromLocalMapId < (int)allData.size())) fromLocalMapPose = allData[fromLocalMapId]->estimatedGlobalPose;
-		if ((toLocalMapId >= 0) || (toLocalMapId < (int)allData.size())) toLocalMapPose = allData[toLocalMapId]->estimatedGlobalPose;
+		if ((fromLocalMapId >= 0) || (fromLocalMapId < allData.size())) fromLocalMapPose = allData[fromLocalMapId]->estimatedGlobalPose;
+		if ((toLocalMapId >= 0) || (toLocalMapId < allData.size())) toLocalMapPose = allData[toLocalMapId]->estimatedGlobalPose;
 		return ORUtils::SE3Pose(toLocalMapPose.GetM() * fromLocalMapPose.GetInvM());
 	}
 }
